E2E Test Cleanup (#5342)
* Remove edu required-only e2e tests

* Rename edu e2e tests

* HCA e2e cleanup

* Rename required -> all-fields

* Actually add that file...

* Rename 5490 all fields test

* Click only one option

* Cleaning up hca tests
