--- test/edu-benefits/5490/00-required.e2e.spec.js
@@ -1,6 +1,0 @@
-const E2eHelpers = require('../../e2e/e2e-helpers');
-const runTest = require('./e2e-test.js');
-
-module.exports = E2eHelpers.createE2eTest(
-  (client) => runTest(client, true)
-);
